backend/requirements.txt
fastapi
uvicorn
python-dotenv
SQLAlchemy
psycopg2-binary
alembic
openpyxl
python-multipart
bcrypt
python-jose[cryptography]
